+++ src/main/java/com/cappleapple/myhudnotyours/model/TextureScaleMode.java
@@ -3,6 +3,7 @@ package com.cappleapple.myhudnotyours.model;
 public enum TextureScaleMode {
     STRETCH,
     TILE,
+    SEGMENTED,
     NINE_SLICE;
 
     public TextureScaleMode next() {